requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
gspread==5.12.4
oauth2client==4.1.3
python-multipart==0.0.6
